requirements.txt
autopep8==2.0.1
chardet==5.1.0
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
protobuf==3.20.3
pycodestyle==2.10.0
python-dotenv==0.21.1
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
